+++ UKS_Git_Site/GitApp/templates/base.html
@@ -19,6 +19,7 @@
         </br>
         <a href="{% url 'logout' %}">Logout</a>
         {% else %}
+        <a href="{% url 'user_register' %}">Register</a>
         <a href="{% url 'login' %}">Login</a>
         {% endif %}
       </div>
