+++ UKS_Git_Site/GitApp/templates/user_form.html
@@ -4,5 +4,4 @@
 
 {% block content %} 
     {%crispy form %}
-    <button onclick="location.href = '/change_password'">Change password</button>
 {% endblock %}
