+++ UKS_Git_Site/GitApp/templates/user_profile.html
@@ -2,11 +2,12 @@
 
 {% block content %}
     <h1>User Details</h1>
-    <h2>First Name: {{object.first_name}}</h2>
+    <p>First Name: {{object.first_name}}</p>
+    <p>Last Name: {{object.last_name}}</p>
     <p>Username: {{ object.username }}</p>
-    {{object.as_p}}
+    <p>Email: {{ object.email }}</p>
     <form action="{% url 'delete_user' %}">
-     <button type="submit">Delete account</button>
+     <button class="btn btn-primary" type="submit">Delete account</button>
     </form>
     <button class="btn btn-primary" onclick="location.href = '/edit_profile'">Edit profile</button>
     <button class="btn btn-primary" onclick="location.href = '/change_password'">Change password</button>
